Create Eilean Siar 2012, Ward 7 data
Data from the Scottish Council Area election Eilean Siar, Ward 7, in 2012.  The first line of data is the number of candidates (10) and the number of seats (4).  After that the data is in the form
(number of ballots) (ranking) 0.  For example, line 4 indicates that 1 ballot had the ranking 1 in first place, 10 in second, 5 in third, and 6 in 4th place.  The key for which number corresponds to which candidates is listed at the bottom of the file.